Move memorytask to tasks namespace package
