+++ research_assistant/tasks/__init__.py
@@ -1,0 +1,1 @@
+"""Namespace package for L'ART Research Assistant Tasks."""
