+++ research_assistant/tasks/memorytask/localisations/GswGer_Ger_DE.json
